requirements.txt
# ===============================================
# SkillSpace 项目后端 Python 依赖
# ===============================================
# 
# 安装说明：
# 1. 创建虚拟环境：python -m venv venv
# 2. 激活虚拟环境：
#    - Windows: venv\Scripts\activate
#    - Linux/Mac: source venv/bin/activate
# 3. 安装依赖：pip install -r requirements.txt
#
# 注意：
# - torch相关包需要CUDA支持，请根据实际GPU情况调整
# - 如果没有GPU，可以安装CPU版本的torch
# - AI模型相关包体积较大，首次安装需要时间
#
# ===============================================

# ================= Django核心框架 =================
Django==4.2.27
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-simpleui==2025.5.262

# ================= 数据库相关 =================
PyMySQL==1.1.2
psutil==7.1.3

# ================= 任务队列（Celery）=================
celery==5.5.3
amqp==5.3.1
billiard==4.2.3
kombu==5.5.4
vine==5.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0

# ================= AI模型核心依赖 =================
# PyTorch（CUDA 12.4版本）
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchvision==0.19.1+cu124

# Transformers生态
transformers==4.57.3
transformers-stream-generator==0.0.4
tokenizers==0.22.1
sentencepiece==0.1.99
safetensors==0.7.0
accelerate==1.10.1
peft==0.10.0

# 量化和优化
bitsandbytes==0.47.0
einops==0.8.1

# 模型下载和管理
huggingface-hub==0.36.0
modelscope==1.11.0

# ================= 数据处理和科学计算 =================
numpy==2.0.2
pandas==2.3.3
scipy==1.13.1
pyarrow==21.0.0

# 数据集处理
datasets==4.4.1
dill==0.4.0
multiprocess==0.70.18
xxhash==3.6.0

# ================= 网络请求和HTTP =================
requests==2.32.5
urllib3==2.5.0
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11

# 异步HTTP
aiohttp==3.13.2
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
async-timeout==5.0.1
frozenlist==1.8.0
multidict==6.7.0
propcache==0.4.1
yarl==1.22.0

# 同步HTTP高级功能
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
anyio==4.12.0

# ================= 阿里云SDK（ModelScope依赖）=================
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
oss2==2.19.1
crcmod==1.7
pycryptodome==3.23.0

# ================= 工具库 =================
# 图像处理
pillow==11.3.0

# 时间处理
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# 环境变量
python-dotenv==1.2.1

# 配置文件
PyYAML==6.0.3
tomli==2.3.0

# 模板引擎
Jinja2==3.1.6
MarkupSafe==3.0.3

# JSON处理
simplejson==3.20.2

# 进度条
tqdm==4.67.1

# 正则表达式增强
regex==2025.11.3

# 命令行交互
prompt_toolkit==3.0.52
wcwidth==0.2.14

# ================= 加密和安全 =================
cryptography==46.0.3
cffi==2.23
pycparser==2.23

# ================= 文件和路径处理 =================
filelock==3.19.1
fsspec==2025.10.0
platformdirs==4.4.0

# ================= 数学和符号计算 =================
sympy==1.14.0
mpmath==1.3.0

# ================= 深度学习工具 =================
networkx==3.2.1
tiktoken==0.7.0
nvidia-ml-py3==7.352.0

# ================= 代码格式化 =================
yapf==0.43.0

# ================= 其他工具 =================
packaging==25.0
attrs==25.4.0
sortedcontainers==2.4.0
exceptiongroup==1.3.1
typing_extensions==4.15.0
six==1.17.0
colorama==0.4.6
addict==2.4.0
gast==0.7.0
jmespath==0.10.0

# ================= Django项目特定 =================
asgiref==3.11.0
sqlparse==0.5.3

# ===============================================
# 安装顺序建议：
# 1. 先安装PyTorch（根据CUDA版本）
# 2. 再安装transformers和相关AI库
# 3. 最后安装其他依赖
#
# 如需CPU版本PyTorch，请访问：
# https://pytorch.org/get-started/locally/
#
# CUDA版本查看命令：nvidia-smi
# ===============================================
